ml/requirements.txt
torch==2.3.0
transformers==4.44.0
peft==0.12.0
trl==0.10.1
bitsandbytes==0.43.3
datasets==2.21.0
accelerate==0.33.0
albumentations==1.4.14
playwright==1.46.0
fastapi==0.112.0
uvicorn==0.30.6
llama-cpp-python==0.2.90
evaluate==0.4.2
rouge-score==0.1.2
nltk==3.8.1
Pillow==10.4.0
httpx==0.27.0
numpy==1.26.4
tqdm==4.66.5
huggingface_hub==0.24.5
safetensors==0.4.4
sentencepiece==0.2.0
protobuf==5.27.2
opencv-python-headless==4.10.0.84
beautifulsoup4==4.12.3
